changed drag-and-drop policy, authentication message, load progress
Changed drag-and-drop policy so that individual files could be loaded
instead of all files in a directory.

Change authentication message to use Level 2 instead of Aria.

Changed GUI so that it updates file summary display on main screen as
files are loaded.  With the old way, the GUI froze until all files were
read in, and users did not know if it was working or hung.