clear anon dest, show preview series, local java
Clear the anonymization destination when all series are cleared.
Indicate on the main dialog which series is being previewed by setting
the text color to green.
Modified bat file to use local version of java if it is available.  This
makes the program start faster because it does not have to load the JVM
over the network.