transfer syntax, anon write permit, dev flag,java
Fixed TransferSyntax problem where ARIA would refuse to import
anonymized files.
Added check in bat file to see if the local installation of java could
be used.  It checks to see if it is version 6 or 7, and if so, uses it.
If not, then it runs an installation from the network, which is usually
much slower on starrtup.
Added better checks and messages to user to handl the problem where the
user does not have permission to write files to the chosen anonymization
destination.  This can happen because of a read only file system or
getting files from a DVD or CD.
Added configuration option for specifying DICOM service URL as a
property on the command line to override the configuration file.  This
was done because as a developer (me - Jim Irrer) often set it to be a
local server on the development machine, and there is the risk of
forgetting to set it back before creating the package and deploying it
to users.