direct DICOM Upload to PACS, threaded GUI
- Files are sent directly from the program to the destination, instead
of going through the
DICOM service.  This is ok for the Conquest, Staging, and Archive PACS,
but may be
a problem if you try to push something directly to Aria, but I don't
think that people do
that very frequently, if ever.

- More functionality is done with threads in the background which
reduces the frequency
of short 'freezes'.

- The user can control whether or not a KO file is uploaded (the
Manifest checkbox).  This
feature was actually on the previous release (1.0.25) but you may not
have used that version.
This feature defaults to always send the extra file, but default can be
changed in the configuration.

The anonymization and preview functionality has not changed.