New MIL version.
Added capability to truncate dates to the year as part of anonymization,
changing both month and day to 01.  This is overridden by values set in
Anonymization Options.

Added -h option for printing help.

Added ability to define application name in configuration file.  This
helps the user know that they are using the intended configuration file.

Anonymization Options window no longer shows a text field for UIDs.  It
was misleading because it inferred that the user could enter a value and
it would be used when in fact it was ignored.