Improved robustness of startup. Fixed compiler errors for new dev env.
When starting up, the java.class.path property is used to get a list of
directories to search for the main and PACS configuration files, but
only after looking in the conventional places.

Moved development environment from Eclipse mars to oxygen.  Some new
compiler errors showed up and had to be fixed.  The compiler now
complains about @Override's.