Added check for command mode before displaying an Alert box.
If an out of memory exception occurred in command line mode, then the
program would try to show an alert box but would have no GUI to display
it to and so would get an exception.  The solution was to check for
command line mode before showing an alert box.