Fixed off-by-4 group length.  Also added fix for handling a rare case of no group length.
The calculation for group length was off by 4 bytes.  It would probably not cause a problem for most software, but could.

It is possible, (but unlikely) that the user would edit a file and remove
the FileMetaInformationGroupLength attribute.  This handles that case by
recreating it before computing the metadata length.